+++ backend/src/main/java/com/carffeine/carffeine/station/domain/jwt/Jwt.java
@@ -1,23 +1,59 @@
 package com.carffeine.carffeine.station.domain.jwt;
 
+import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 import java.security.Key;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.Date;
 
 @Getter
-@RequiredArgsConstructor
+@Component
+@NoArgsConstructor
 public class Jwt {
 
-    private final String token;
     @Value("${jwt.secret}")
     private String secret;
 
-    public boolean isExpired() {
+    public String createJwt(Long id) {
+        Claims claims = Jwts.claims();
+        claims.put("id", id);
+        return createToken(claims);
+    }
+
+    private String createToken(Claims claims) {
+        Key key = gethmacShaKey();
+
+        return Jwts.builder()
+                .setClaims(claims)
+                .setIssuedAt(getIssuedAt())
+                .setExpiration(getExpiration())
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    private Key gethmacShaKey() {
+        return Keys.hmacShaKeyFor(secret.getBytes());
+    }
+
+    private Date getIssuedAt() {
+        LocalDateTime now = LocalDateTime.now();
+        return Date.from(now.atZone(ZoneId.systemDefault()).toInstant());
+    }
+
+    private Date getExpiration() {
+        LocalDateTime now = LocalDateTime.now();
+        return Date.from(now.plusHours(1).atZone(ZoneId.systemDefault()).toInstant());
+    }
+
+    public boolean isExpired(String token) {
         return Jwts.parser()
                 .setSigningKey(secret)
                 .parseClaimsJws(token)
@@ -26,15 +62,11 @@ public class Jwt {
                 .before(new Date());
     }
 
-    public Long extractId() {
+    public Long extractId(String token) {
         return Jwts.parser()
                 .setSigningKey(secret)
                 .parseClaimsJws(token)
                 .getBody()
                 .get("id", Long.class);
-    }
-
-    public Key gethmacShaKey() {
-        return Keys.hmacShaKeyFor(secret.getBytes());
     }
 }
