+++ backend/src/main/java/com/carffeine/carffeine/station/service/auth/MemberService.java
@@ -1,18 +1,16 @@
 package com.carffeine.carffeine.station.service.auth;
 
 import com.carffeine.carffeine.station.domain.jwt.Jwt;
-import com.carffeine.carffeine.station.domain.jwt.JwtProvider;
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
 @Service
 public class MemberService {
 
-    private final JwtProvider jwtProvider;
+    @Autowired
+    private Jwt jwt;
 
     public String createJwt(Long memberId) {
-        Jwt jwt = jwtProvider.createJwt(memberId);
-        return jwt.getToken();
+        return jwt.createJwt(memberId);
     }
 }
